requirements.txt
flake8
flask
mwapi
pytest
mypy
pyyaml
toolforge
